feat: improve improve/exclude (#44)
Reality of the situation in Deno:

1. `"exclude"` has higher precedence than `"include"`.
   - A file in `"include"` will not be matched against if it's excluded.
2. `"include"` is a closed set.
   - Files not in the include set are not included.

This change makes it more flexible by having the ability to opt out of
excluded files. This can be achieved via negated globs in either
"publish.exclude" or "exclude" (or "fmt.exclude", "exclude" for the fmt
command).

Specifically for checking if a file is excluded for publishing we first
check `"publish.include"` for a match, then:

1. We use the array of the config's "exclude" + "publish.exclude" (done
in https://github.com/denoland/deno_config/pull/43)
2. Go through each entry in reverse.
1. If entry is matched as opted-in (ex. !/dist/), then we consider that
as NOT excluded.
2. If entry is matched for an exclusion (ex. /dist/file.txt), then we
consider that as excluded.
   3. We bail on the first positive or negative match for perf reasons.
3. At this point, if the file has not had a positive or negative match,
we check if the gitignore excludes the file.