fix: discover tsconfig.json if there's a deno.json too, only discover at root, and only if `compilerOptions` not set (#168)
This changes it to also discover a tsconfig.json if there's a deno.json
too, but the deno.json must not contain `compilerOptions`.

Additionally, only discovers the root tsconfig.json for now. I think we
need to think through some edge cases before enabling it in member
directories of a workspace. For example, what occurs if the root
deno.json has `compilerOptions` set and a tsconfig.json and a member
directory only has a tsconfig.json?